+++ Aoc2024/Days/Day04.cs
@@ -33,7 +33,7 @@ public sealed class Day04 : IDay
         """;
 
     [Example(TestInput1, 18)]
-    public int Part1(IInput input)
+    public long Part1(IInput input)
     {
         var m = Matrix.Parse(input.Text);
 
@@ -71,7 +71,7 @@ public sealed class Day04 : IDay
     }
 
     [Example(TestInput2, 9)]
-    public int Part2(IInput input)
+    public long Part2(IInput input)
     {
         var m = Matrix.Parse(input.Text);
         var result = 0;
