+++ Aoc2024/Lib/IDay.cs
@@ -2,7 +2,7 @@ namespace Aoc2024.Lib;
 
 public interface IDay
 {
-    int Part1(IInput input);
+    long Part1(IInput input);
 
-    int Part2(IInput input);
+    long Part2(IInput input);
 }
